Move `dns-control` to `bin-shared
