+++ rust/bin-shared/Cargo.toml
@@ -10,6 +10,7 @@ license = { workspace = true }
 anyhow = { workspace = true }
 axum = { workspace = true, features = ["http1", "tokio"] }
 clap = { workspace = true, features = ["derive", "env"] }
+dns-types = { workspace = true }
 firezone-logging = { workspace = true }
 futures = { workspace = true, features = ["std", "async-await"] }
 gat-lending-iterator = { workspace = true }
@@ -18,7 +19,7 @@ ip-packet = { workspace = true }
 ip_network = { workspace = true, features = ["serde"] }
 socket-factory = { workspace = true }
 thiserror = { workspace = true }
-tokio = { workspace = true, features = ["io-util", "net", "rt", "sync"] }
+tokio = { workspace = true, features = ["io-util", "net", "rt", "sync", "process"] }
 tracing = { workspace = true }
 tun = { workspace = true }
 
@@ -29,11 +30,13 @@ tokio = { workspace = true, features = ["macros", "rt-multi-thread"] }
 
 [target.'cfg(target_os = "linux")'.dependencies]
 dirs = { workspace = true }
+atomicwrites = { workspace = true }
 flume = { workspace = true }
 libc = { workspace = true }
 netlink-packet-core = { version = "0.7" }
 netlink-packet-route = { version = "0.19" }
 nix = { workspace = true, features = ["socket"] }
+resolv-conf = { workspace = true }
 rtnetlink = { workspace = true }
 zbus = { workspace = true } # Can't use `zbus`'s `tokio` feature here, or it will break toast popups all the way over in `gui-client`.
 
@@ -46,6 +49,8 @@ windows-implement = { workspace = true }
 wintun = "0.5.1"
 winreg = { workspace = true }
 tokio-util = { workspace = true }
+ipconfig = "0.3.2"
+itertools = { workspace = true }
 
 [target.'cfg(windows)'.dependencies.windows]
 workspace = true
@@ -68,6 +73,10 @@ features = [
 [target.'cfg(windows)'.dev-dependencies]
 ip-packet = { workspace = true }
 tokio = { workspace = true, features = ["net", "time"] }
+
+[target.'cfg(target_os = "linux")'.dev-dependencies]
+mutants = "0.0.3" # Needed to mark functions as exempt from `cargo-mutants` testing
+tempfile = { workspace = true }
 
 [lints]
 workspace = true
