+++ rust/bin-shared/tests/network_notifiers.rs
@@ -1,6 +1,6 @@
 #![allow(clippy::unwrap_used)]
 
-use firezone_bin_shared::{new_dns_notifier, new_network_notifier, platform::DnsControlMethod};
+use firezone_bin_shared::{DnsControlMethod, new_dns_notifier, new_network_notifier};
 use futures::future::FutureExt as _;
 use std::time::Duration;
 use tokio::time::timeout;
